ingestion/requirements.txt
functions-framework==3.*
requests==2.31.0
google-cloud-storage==2.14.0
google-cloud-pubsub==2.19.0
google-cloud-secret-manager==2.18.0
google-cloud-firestore==2.14.0